converted original tick date file to append mock SSS tick data file

+++ src/utils/MathUtil.java
@@ -35,14 +35,14 @@ public class MathUtil {
 	 * @param totalInts
 	 * @return
 	 */
-	public static List<Integer> getUniqueRandomIntSet(int start, int end, Random random, int totalInts) {
-		List<Integer> wholeList = new ArrayList<Integer>();
+	public static List<Long> getUniqueRandomIntSet(long start, long end, Random random, int totalInts) {
+		List<Long> wholeList = new ArrayList<Long>();
 		
-		for(int i=start; i<=end; i++){
+		for(long i=start; i<=end; i++){
 			wholeList.add(i);
 		}
 		Collections.shuffle(wholeList, random);
-		List<Integer> resultList = wholeList.subList(0,totalInts);
+		List<Long> resultList = wholeList.subList(0,totalInts);
 		Collections.sort(resultList);
 		return resultList;
 	}
